Move npk crate into northstar module
