+++ main/Cargo.toml
@@ -27,4 +27,4 @@ android-logd-logger = "0.1.3"
 
 [features]
 default = []
-hello-world = ["northstar/hello-world"]
+#hello-world = ["northstar/hello-world"]
