+++ northstar/src/runtime/mount.rs
@@ -20,12 +20,12 @@ use super::{
     loopdev::LoopControl,
     state::Npk,
 };
+use crate::{npk, npk::dm_verity::VerityHeader};
 use bitflags::_core::str::Utf8Error;
 use floating_duration::TimeAsFloat;
 use futures::{future::ready, Future, FutureExt};
 use log::{debug, info};
 pub use nix::mount::MsFlags as MountFlags;
-use npk::dm_verity::VerityHeader;
 use std::{
     io,
     os::unix::io::AsRawFd,
