--- npk/Cargo.toml
@@ -1,33 +1,0 @@
-[package]
-name = "npk"
-version = "0.1.2"
-authors = ["ESRLabs"]
-edition = "2018"
-
-[dependencies]
-base64 = "0.13.0"
-byteorder = "1.4"
-caps = "0.5.1"
-ed25519-dalek = "1.0"
-fs_extra = "1.2"
-hex = "0.4.3"
-humanize-rs = "0.1.5"
-itertools = "0.10.0"
-log = { version = "0.4.14", features = ["serde"] }
-rand = "0.7.3"
-regex = "1.4"
-semver = { version = "0.11.0", features = ["serde"] }
-serde = { version = "1.0", features = ["derive"] }
-serde_with = "1.9"
-serde_yaml = "0.8.17"
-sha2 = "0.9.3"
-tempfile = "3.2"
-thiserror = "1.0"
-uuid = { version = "0.8.2", features = ["v4"] }
-which = "4.1"
-zip = { version = "0.5.13", features = ["unreserved"] }
-
-[dev-dependencies]
-anyhow = "1.0"
-tempfile  = "3.2"
-tokio = { version = "1.5", features = ["full"] }
