+++ tools/sextant/Cargo.toml
@@ -11,7 +11,7 @@ colored = "2.0"
 ed25519-dalek = "1.0"
 env_logger = "0.8.3"
 log = "0.4.14"
-npk = { path = "../../npk" }
+northstar = { path = "../../northstar"}
 structopt = "0.3.21"
 tempfile = "3.2"
 which = "4.1"
